pyproject.toml
[project]
name = "cognee-qdrant-starter"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "python-dotenv>=1.0",
    "boto3>=1.35",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"
